removing old GameRunner and JUnit Vintage
we don't use the annotation parts of spf4j-slf4j-test enough to keep Vintage around
